requirements.txt
streamlit
isodate
google-api-python-client 
pandas 
transformers 
torch 
plotly-express
scikit-learn